--- src/main/kotlin/auth/dto/AuthToken.kt
@@ -1,11 +1,0 @@
-package uket.auth.dto
-
-data class AuthToken(
-    val accessToken: String,
-    val refreshToken: String,
-    val isRegistered: Boolean,
-) {
-    companion object {
-        fun of(accessToken: String, refreshToken: String, isRegistered: Boolean): AuthToken = AuthToken(accessToken, refreshToken, isRegistered)
-    }
-}
