+++ src/main/kotlin/auth/dto/UserAuthToken.kt
@@ -1,0 +1,11 @@
+package uket.auth.dto
+
+data class UserAuthToken(
+    val accessToken: String,
+    val refreshToken: String,
+    val isRegistered: Boolean,
+) {
+    companion object {
+        fun of(accessToken: String, refreshToken: String, isRegistered: Boolean): UserAuthToken = UserAuthToken(accessToken, refreshToken, isRegistered)
+    }
+}
