+++ src/main/kotlin/auth/util/AuthTokenGenerator.kt
@@ -1,7 +1,7 @@
-package uket.domain.user.service
+package uket.auth.util
 
 import org.springframework.stereotype.Component
-import uket.auth.dto.AuthToken
+import uket.auth.dto.UserAuthToken
 import uket.auth.jwt.JwtAuthTokenUtil
 import uket.domain.user.entity.User
 import uket.domain.user.enums.UserRole
@@ -10,7 +10,7 @@ import uket.domain.user.enums.UserRole
 class AuthTokenGenerator(
     private val jwtAuthTokenUtil: JwtAuthTokenUtil,
 ) {
-    fun generateAuthToken(user: User): AuthToken {
+    fun generateAuthToken(user: User): UserAuthToken {
         val userId: Long = user.id
         val name: String = user.name
         val isRegistered: Boolean = user.isRegistered
@@ -18,6 +18,6 @@ class AuthTokenGenerator(
         val newAccessToken: String = jwtAuthTokenUtil.createAccessToken(userId, name, UserRole.USERS.toString(), isRegistered)
         val newRefreshToken: String = jwtAuthTokenUtil.createRefreshToken()
 
-        return AuthToken.of(newAccessToken, newRefreshToken, isRegistered)
+        return UserAuthToken.of(newAccessToken, newRefreshToken, isRegistered)
     }
 }
