+++ src/main/kotlin/facade/UserAuthFacade.kt
@@ -1,21 +1,23 @@
-package uket.domain.user.service
+package uket.facade
 
 import io.lettuce.core.RedisException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import uket.auth.dto.AuthToken
+import uket.auth.dto.UserAuthToken
 import uket.auth.dto.response.userinfo.OAuth2UserInfoResponse
 import uket.auth.filter.TokenValidator
 import uket.auth.jwt.JwtValues.JWT_PAYLOAD_VALUE_REFRESH
 import uket.auth.oauth.OAuth2TokenManager
+import uket.auth.util.AuthTokenGenerator
 import uket.auth.util.OAuth2UserInfoManager
 import uket.domain.user.dto.CreateUserCommand
 import uket.domain.user.entity.User
 import uket.domain.user.enums.Platform
+import uket.domain.user.service.UserService
 import uket.modules.redis.service.RotateTokenService
 
 @Service
-class AuthService(
+class UserAuthFacade(
     private val tokenValidator: TokenValidator,
     private val oauth2TokenManager: OAuth2TokenManager,
     private val oAuth2UserInfoManager: OAuth2UserInfoManager,
@@ -24,19 +26,19 @@ class AuthService(
     private val rotateTokenService: RotateTokenService,
 ) {
     @Transactional
-    fun login(platform: Platform, redirectUri: String, code: String): AuthToken {
+    fun login(platform: Platform, redirectUri: String, code: String): UserAuthToken {
         val tokenResponse = oauth2TokenManager.getAccessToken(platform, redirectUri, code)
         val userInfo = oAuth2UserInfoManager.getUserInfo(platform, tokenResponse)
 
         val newUser = userService.createUser(generateCreateUserDto(userInfo))
 
-        val authToken: AuthToken = authTokenGenerator.generateAuthToken(newUser)
+        val userAuthToken: UserAuthToken = authTokenGenerator.generateAuthToken(newUser)
 
-        rotateTokenService.storeToken(authToken.refreshToken, authToken.accessToken, newUser.id)
-        return authToken
+        rotateTokenService.storeToken(userAuthToken.refreshToken, userAuthToken.accessToken, newUser.id)
+        return userAuthToken
     }
 
-    fun reissue(accessToken: String, refreshToken: String): AuthToken {
+    fun reissue(accessToken: String, refreshToken: String): UserAuthToken {
         tokenValidator.checkNotExpiredToken(accessToken)
         val existingAccessToken: String? = rotateTokenService.getAccessTokenForToken(refreshToken)
         if (accessToken != existingAccessToken) {
@@ -51,10 +53,10 @@ class AuthService(
 
         val findUser: User = userService.getById(rotateTokenService.getUserIdForToken(refreshToken))
 
-        val authToken: AuthToken = authTokenGenerator.generateAuthToken(findUser)
+        val userAuthToken: UserAuthToken = authTokenGenerator.generateAuthToken(findUser)
 
-        rotateTokenService.storeToken(authToken.refreshToken, authToken.accessToken, findUser.id)
-        return authToken
+        rotateTokenService.storeToken(userAuthToken.refreshToken, userAuthToken.accessToken, findUser.id)
+        return userAuthToken
     }
 
     private fun generateCreateUserDto(userInfo: OAuth2UserInfoResponse): CreateUserCommand = CreateUserCommand(
